cannon: Move instrumented_state logic into singlethreaded pkg
